Adding one fix to segment_identity_daily. Delete unused transform/read_source.py. Move used ReadSource class out of segment_identity/transforms.py to its own file. Set up to incorporate fragments into segment_identity_daily.
